requirements.txt
feather-format==0.4.0
joblib==0.14.1
nltk==3.4.5
numpy==1.18.1
pandas==1.0.1
pyarrow==0.16.0
python-dateutil==2.8.1
pytz==2019.3
regex==2020.2.20
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
